:axe: separate styles
